move upload to filesystem_run
